Revert "Feature/#39 Make Calcute & UI update"
