+++ lib/main.dart
@@ -11,9 +11,9 @@ import 'package:damyo/screens/home/map/somking_area/review/write_review_screen.d
 import 'package:damyo/screens/home/map/somking_area/smoking_area_info_screen.dart';
 import 'package:damyo/screens/home/mypage/in_mypage/favorite_screen.dart';
 import 'package:damyo/screens/home/statistics/local_statistics.dart';
-import 'package:damyo/screens/home/statistics/periodsingle_statistics.dart';
+import 'package:damyo/screens/home/statistics/period_statistics.dart';
 import 'package:damyo/screens/home/statistics/time_statistics.dart';
-import 'package:damyo/screens/home/statistics/periodcompare_statistics.dart';
+import 'package:damyo/screens/home/statistics/week_statistics.dart';
 import 'package:damyo/screens/login/login_screen.dart';
 import 'package:damyo/screens/home/home_screen.dart';
 import 'package:damyo/screens/signup/signup_screen.dart';
@@ -195,6 +195,42 @@ final GoRouter router = GoRouter(
             final title = (state.extra as Map<String, String>)['title'] ??
                 "Default Title";
             return ChallengeVoteScreen(title: title);
+          },
+        ),
+        GoRoute(
+          name: 'local_statistics',
+          path: 'local_statistics',
+          builder: (context, state) {
+            return LocalStaisticsScreen(
+              subcategory: state.extra.toString(),
+            );
+          },
+        ),
+        GoRoute(
+          name: 'time_statistics',
+          path: 'time_statistics',
+          builder: (context, state) {
+            return TimeStaisticsScreen(
+              subcategory: state.extra.toString(),
+            );
+          },
+        ),
+        GoRoute(
+          name: 'period_statistics',
+          path: 'period_statistics',
+          builder: (context, state) {
+            return PeriodStaisticsScreen(
+              subcategory: state.extra.toString(),
+            );
+          },
+        ),
+        GoRoute(
+          name: 'week_statistics',
+          path: 'week_statistics',
+          builder: (context, state) {
+            return WeekStaisticsScreen(
+              subcategory: state.extra.toString(),
+            );
           },
         ),
       ],
