+++ lib/screens/home/statistics/statistics_screen.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 import 'package:damyo/http.dart';
@@ -17,7 +15,7 @@ class _StatisticsScreenState extends State<StatisticsScreen> {
   @override
   Widget build(BuildContext context) {
     return ScreenUtilInit(
-      designSize: const Size(390, 1863),
+      designSize: const Size(double.infinity, 1863),
       builder: (context, child) => Scaffold(
         appBar: AppBar(
           scrolledUnderElevation: 0,
@@ -47,7 +45,8 @@ class _StatisticsScreenState extends State<StatisticsScreen> {
                 ),
                 // 지역별 흡연 데이터 통계
                 SizedBox(
-                  height: 600.h,
+                  width: 390,
+                  height: 500.h,
                   child: PageView(
                     scrollDirection: Axis.horizontal,
                     children: const [
@@ -70,47 +69,51 @@ class _StatisticsScreenState extends State<StatisticsScreen> {
                     ],
                   ),
                 ),
-                Container(
+                SizedBox(
+                  width: 390.w,
                   height: 20.h,
-                  color: Colors.grey[200],
-                ),
-                const SizedBox(
-                  child: StatistTap(
-                    index: 6,
-                    category: '특정 기간',
-                  ),
-                ),
-                Container(
-                  height: 20.h,
-                  color: Colors.grey[200],
                 ),
                 // 시간대별 흡연 데이터 통계(사용자, 전체)
                 SizedBox(
-                  height: 1000.h,
-                  child: const StatistTap(
-                    index: 2,
-                    category: '시간대별 평균 흡연량',
+                  width: 390,
+                  height: 469,
+                  child: PageView(
+                    scrollDirection: Axis.horizontal,
+                    children: const [
+                      StatistTap(
+                        index: 2,
+                        category: '시간대별 평균 흡연량',
+                      ),
+                      // StatistTap(index: 1, category: '시간대별 평균 흡연량(전체)',),
+                    ],
                   ),
                 ),
-                Container(
+                SizedBox(
+                  width: 390.w,
                   height: 20.h,
-                  color: Colors.grey[200],
                 ),
                 // 기간별 흡연 데이터 통계(사용자)
                 SizedBox(
-                  height: 1000.h,
-                  child: const StatistTap(
-                    index: 3,
-                    category: '기간별 평균 흡연량(개인)',
+                  width: 390,
+                  height: 469,
+                  child: PageView(
+                    scrollDirection: Axis.horizontal,
+                    children: const [
+                      StatistTap(
+                        index: 3,
+                        category: '기간별 평균 흡연량(개인)',
+                      ),
+                    ],
                   ),
                 ),
-                Container(
+                SizedBox(
+                  width: 390.w,
                   height: 20.h,
-                  color: Colors.grey[200],
                 ),
                 // 담배값 계산기
                 SizedBox(
-                  height: 500.h,
+                  width: 390,
+                  height: 469,
                   child: PageView(
                     scrollDirection: Axis.horizontal,
                     children: const [
@@ -121,16 +124,22 @@ class _StatisticsScreenState extends State<StatisticsScreen> {
                     ],
                   ),
                 ),
-                Container(
+                SizedBox(
+                  width: 390.w,
                   height: 20.h,
-                  color: Colors.grey[200],
                 ),
                 // 평균 흡연량 비교 데이터 통계
                 SizedBox(
-                  height: 1000.h,
-                  child: const StatistTap(
-                    index: 5,
-                    category: '평균 흡연량 비교 통계',
+                  width: 390,
+                  height: 469,
+                  child: PageView(
+                    scrollDirection: Axis.horizontal,
+                    children: const [
+                      StatistTap(
+                        index: 5,
+                        category: '평균 흡연량 비교 통계',
+                      ),
+                    ],
                   ),
                 ),
                 SizedBox(
