+++ pubspec.yaml
@@ -51,8 +51,7 @@ dependencies:
   share_plus: ^8.0.3
   fl_chart: ^0.67.0
   fluttertoast: ^8.2.5
-  intl: ^0.17.0
-  
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
